requirements.txt
beautifulsoup4==4.10.0
certifi==2022.9.24
charset-normalizer==2.0.12
click==8.1.3
Flask==2.2.2
greenlet==1.1.3.post0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jqdatasdk==1.8.11
MarkupSafe==2.1.1
msgpack==1.0.4
numpy==1.23.4
pandas==1.3.5
ply==3.11
pyeastmoney==0.1.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.5
requests==2.30.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==1.4.42
thriftpy2==0.4.14
tqdm==4.65.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.10.0
